// sanity.types.ts
export type Post = {
    _id: string;
    _type: "post";
    _createdAt: string;
    _updatedAt: string;
    _rev: string;
    title?: string;
    slug?: Slug;
    author?: {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "author";
    };
    mainImage?: {
        asset?: {
            _ref: string;
            _type: "reference";
            _weak?: boolean;
            [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
        };
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        alt?: string;
        _type: "image";
    };
    blogCategories?: Array<{
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        _key: string;
        [internalGroqTypeReferenceTo]?: "blogCategory";
    }>;
    publishedAt?: string;
    body?: Array<
        | {
              children?: Array<{
                  marks?: Array<string>;
                  text?: string;
                  _type: "span";
                  _key: string;
              }>;
              style?: "normal" | "h1" | "h2" | "h3" | "h4" | "blockquote";
              listItem?: "bullet";
              markDefs?: Array<{
                  href?: string;
                  _type: "link";
                  _key: string;
              }>;
              level?: number;
              _type: "block";
              _key: string;
          }
        | {
              asset?: {
                  _ref: string;
                  _type: "reference";
                  _weak?: boolean;
                  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
              };
              hotspot?: SanityImageHotspot;
              crop?: SanityImageCrop;
              alt?: string;
              _type: "image";
              _key: string;
          }
    >;
};
// ...
export type Slug = {
    _type: "slug";
    current?: string;
    source?: string;
};
// ...
export type SanityImageCrop = {
    _type: "sanity.imageCrop";
    top?: number;
    bottom?: number;
    left?: number;
    right?: number;
};

export type SanityImageHotspot = {
    _type: "sanity.imageHotspot";
    x?: number;
    y?: number;
    height?: number;
    width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;